Move the 'mappertype' to job configuration.
This means that scripts do not have to implement an extra method to set the
mapper type, and it can just be set once in that data source's "setupjob"
function. This eases the burden for developing new scripts.

Any scripts already written to specify a mappertype function should continue
to work fine, as long as they reference one of the standard jydoop 'setupjob'
functions.